+++ .eslintrc.js
@@ -1,16 +1,10 @@
+const evernautEslintConfigBase = require("./src/index");
+
 module.exports = {
+  ...evernautEslintConfigBase,
   env: {
     es6: true,
     node: true,
   },
-  extends: ["prettier"],
-  parserOptions: {
-    ecmaVersion: 2018,
-    sourceType: "module",
-  },
-  plugins: ["json", "prettier"],
   root: true,
-  rules: {
-    "json/*": ["error"],
-  },
 };