+++ src/eslint-native-rules-disabled-for-spec-files.js
@@ -1,0 +1,5 @@
+module.exports = {
+  "max-lines": "off",
+  "max-params": "off",
+  "max-statements": "off",
+};